requirements.txt
Flask>=2.3
Flask-Cors>=4.0
python-dotenv>=1.0

